requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pandasql>=0.7.3
openai>=0.28.0
python-dotenv>=1.0.0 